major refactor -- make Shed.Blob module hierarchy where new types of blobs exist
Each type can expose a way to index it, a way to render it, a way to import it,
etc. There aren't generic interfaces (as the dispatch would actually get more
complicated), but things are re-organized to make it much easier to add a new
module and then put the various functions in the right place (and the places
where they get put already have lists of such functions set up).