Add delete blob type
This claims a deletion. Provided that the blob being deleted is not itself a
delete blob, this is concurrency safe (i.e., different clients can do different
things, and we can still do the same trivial merging of content and re-apply the
deletion). Given that the delete nodes are tiny and never shown in the UI, they
should never need to be removed (unless you have concern about leaking
information, though all it leaks is the sha of a node that used to be there,
which isn't much).

In the UI, we add buttons on files which create deletions on the file blob and
all the associated content blobs.