Add support for replace blob type
Currently, this only has well-defined semantics in single-instance contexts, as
different instances could replace the same blob with different blobs. However,
the plan is to make this create a landing-page where the user can pick where to
go (so it will be well-defined).

Right now, replace blobs can only be created manually, and in general, they will
not likely be created by users, but show up more as data is mutated (to maintain
stable URLs).