Add support for multiple replacements, with landing pages, and transitive replacements
Note that there is still a little bit of perhaps undesirable behavior: there is
no way of directly deleting a replacement; your only recourse is to delete what
is being pointed to. Given that my expectation is that the primary point of this
mechanism is for (mutable) collections, I don't think this is really a big deal
-- as you essentially just have to do manual conflict resolution and delete the
version you don't want -- but it's perhaps not ideal.