Add basic support for boxes: once created manually, upload/remove works
Files are added to them by opening them up and then uploading to that page.
Deletion has been removed from the UI, as it shouldn't be a common thing and
until there is better UI in general, I want to reserve that for removing from a
box, which will be common.