Add a CachingMemory blob store that layers on top of any other store
This is used by the application by default when using S3.
It's possible that we'll want a persistent cache (i.e., to postgres, redis,
etc), but this is the simplest thing to start with; this means that the cache is
wiped out every time the server is restarted, even though the cache should
always be correct.